LegionAnime : fix in some extractors (#1064)
