+++ src/es/legionanime/build.gradle
@@ -5,7 +5,7 @@ ext {
     extName = 'LegionAnime'
     pkgNameSuffix = 'es.legionanime'
     extClass = '.LegionAnime'
-    extVersionCode = 3
+    extVersionCode = 4
     libVersion = '13'
 }
