+++ src/es/legionanime/src/eu/kanade/tachiyomi/animeextension/es/legionanime/extractors/UploadExtractor.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.animeextension.es.monoschinos.extractors
+package eu.kanade.tachiyomi.animeextension.es.legionanime.extractors
 
 import eu.kanade.tachiyomi.animesource.model.Video
 import eu.kanade.tachiyomi.network.GET
